adds docs and moves 'validatePublicKey' to security/security.go
