+++ pkg/artifactory/resource/security/resource_artifactory_trusted_key.go
@@ -51,12 +51,12 @@ func ResourceArtifactoryTrustedKey() *schema.Resource {
 				Description: "Generated by artifactory on upload",
 			},
 			"public_key": {
-				Type:        schema.TypeString,
-				Required:    true,
-				Description: "The Public key to add as trusted distribution GPG key",
-				ForceNew:    true,
-				//				StateFunc:        stripTabs,
-				ValidateDiagFunc: validatePublicKey, // from resource_artifactory_keypair.go
+				Type:             schema.TypeString,
+				Required:         true,
+				Description:      "The Public key to add as trusted distribution GPG key",
+				ForceNew:         true,
+				StateFunc:        stripTabs,
+				ValidateDiagFunc: validatePublicKey,
 			},
 			"issued_on": {
 				Type:        schema.TypeString,
@@ -126,6 +126,8 @@ func readTrustedKey(_ context.Context, d *schema.ResourceData, m interface{}) di
 		if key.KeyID == d.Id() {
 			d.Set("key_id", key.KeyID)
 			d.Set("fingerprint", key.Fingerprint)
+			// Import functionality still has to be implemented
+			// d.Set("public_key", key.PublicKey)
 			d.Set("issued_on", key.IssuedOn)
 			d.Set("issued_by", key.IssuedBy)
 			d.Set("valid_until", key.ValidUntil)
